chatbot/requirements.txt
openai
python-dotenv
slack-bolt
slack-sdk